jack: Use atomic access for hardware pointer and xrun flag
volatile is not sufficient for guaranteeing the order of buffer access
and hardware pointer update.
In playback it could happen that the hardware pointer is updated but the
audio data is not yet forwarded to the JACK daemon.
Therefore the audio data could be overwritten by the ALSA application
before it was copied to the JACK daemon.

Signed-off-by: Timo Wischer <twischer@de.adit-jv.com>